refactor: reorganize imports and add new model classes and constants
- se añadieron .py segun cada clase en su carpeta correspondiente
- se ajustaron los codigos de src/ y tests/ al estandar PEP8
- se unificaron los imports de todos los archivos .py de src
