+++ src/controllers/controllers.py
@@ -1,0 +1,11 @@
+from src.controllers.ShipmentManager import ShipmentManager
+from src.controllers.PackagesManager import PackageManager
+from src.controllers.LoginManager import LoginManager
+from src.controllers.PaymentsManager import PaymentsManager
+from src.controllers.UsersManager import UsersManager
+
+__all__ = ["ShipmentManager",
+           "PackageManager",
+           "LoginManager",
+           "PaymentsManager",
+           "UsersManager"]
