+++ src/models/models.py
@@ -1,0 +1,6 @@
+from src.models.Shipment import Shipment
+from src.models.Package import Package
+from src.models.User import User
+from src.models.Invoice import Invoice
+
+__all__ = ["Shipment", "Package", "User", "Invoice"]
