+++ src/utils/constants.py
@@ -1,0 +1,5 @@
+NO_FOUND = "NO_FOUND"
+LOGIN_OPTION = "1"
+REGISTER_OPTION = "2"
+EXIT_OPTION = "q"
+MESSAGGE_MAIN_MENU = "MESSAGGE_MAIN_MENU"
