+++ src/views/views.py
@@ -1,0 +1,13 @@
+from src.views.invoices_menu import InvoicesMenu
+from src.views.main_menu import MainMenu
+from src.views.package_menu import PackageMenu
+from src.views.shipment_menu import ShipmentMenu
+from src.views.login_menu import LoginMenu
+from src.views.menu import Menu
+
+__all__ = ["InvoicesMenu",
+           "MainMenu",
+           "PackageMenu",
+           "ShipmentMenu",
+           "LoginMenu",
+           "Menu"]
